restructure files directory
