+++ lib/views/note/note.dart
