--- lib/views/screens/trial.html
@@ -1,40 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Document</title>
-</head>
-<body>
-  <ul>
-    <li><b>Microorganisms:</b>
-      <ul>
-        <li>Small living organisms not visible to the naked eye</li>
-        <li>Wide variety of species</li>
-        <li>Play important roles in various ecosystems</li>
-      </ul>
-    </li>
-    <li><b>Types:</b>
-      <ul>
-        <li>Bacteria:</li>
-        <ul>
-          <li>Single-celled prokaryotes</li>
-          <li>Found virtually everywhere on Earth</li>
-          <li>Some species can be harmful (pathogenic bacteria)</li>
-        </ul>
-        <li>Viruses:</li>
-        <ul>
-          <li>Obligate <mark>intracellular</mark> parasites</li>
-          <li>Consist of genetic material enclosed in a protein coat</li>
-          <li>Cannot reproduce without a host</li>
-        </ul>
-        <li>Fungi:</li>
-        <ul>
-          <li>Eukaryotic organisms</li>
-          <li>Obtain nutrients by absorbing them from their environment</li>
-          <li>Include yeasts, molds, and mushrooms</li>
-        </ul>
-        
-</body>
-</html>
-
